+++ earthmover/operations/column.py
@@ -105,6 +105,11 @@ class ModifyColumnsOperation(Operation):
             raise
 
         # TODO: Allow user to specify string that represents current column value.
+        if 'value' in data.columns:
+            # prevent existing `value` column from being clobbered, if it exists
+            self.error_handler.throw(
+                f"error in `modify_columns` operation; a column named `value` already exists, and would be removed by this operation... please rename it before using `modify_columns`"
+            )
         data['value'] = data[col]
 
         data[col] = data.apply(
@@ -176,7 +181,7 @@ class RenameColumnsOperation(Operation):
 
         for old_col, new_col in self.columns_dict.items():
             if new_col in data.columns:
-                self.logger.warning(
+                self.error_handler.throw(
                     f"Rename column operation overwrites existing column `{new_col}`."
                 )
             if old_col not in data.columns: